refactor new-FamilyShare page into corresponding modules
